finishes refactor of modal, pin form
