+++ src/components/common/LoadingComponent.js
@@ -1,18 +1,15 @@
 import React from 'react';
-// spinner size is controlled in the theme-overrides.js file
 
-// components
-import { Spin } from 'antd';
-
-function LoadingComponent() {
+const LoadingComponent = () => {
   return (
     <div data-testid="spinnerCont" className="loader">
-      {/* takes an optional "tip" prop as text to display under the spinner: (tip= 'loading...') 
-      size options: "small", "default", "large" */}
-      {/* <Spin data-testid="antSpinner" size="large" /> */}
-      <img src="/images/shark-spinner.gif" data-testid="antSpinner" alt="spinnerShark" />
+      <img
+        src="/images/shark-spinner.gif"
+        data-testid="spinnerShark"
+        alt="spinnerShark"
+      />
     </div>
   );
-}
+};
 
 export default LoadingComponent;
