+++ src/components/common/index.js
@@ -1,6 +1,5 @@
 import LoadingComponent from './LoadingComponent';
 import Header from './Header';
-import AccountPinModal from './accountPinModal/AccountPinModal';
 // notice we're building out a 'package' of reusables here and exporting them as an object of component properties.
 // to use this, simply `import {foo, bar, baz} from '<path-to-this-directory>/ReusableComponents';`
-export { LoadingComponent, Header, AccountPinModal };
+export { LoadingComponent, Header };
