+++ src/styles/UserForm.less
@@ -26,7 +26,7 @@
 
   .userText {
     text-align: center;
-    margin: 2%;
+    margin: 4% 2% 2% 2%;
     font-family: 'Nunito', sans-serif;
     font-style: normal;
     font-weight: normal;
